Add some Error messages This will add error messages for when files are missing in certain scenarioes.

+++ src/main.rb
@@ -2,17 +2,21 @@ require 'time'
 
 def startScreen
     if File.exist?("pass")
+        if !File.exist?("settings")
+            puts "ERROR 001"
+            puts "Crucial file missing! Cannot continue"
+            exit(1)
+        end
+
         settings = File.open("settings", "r+")
         settings.rewind
-    
-        masterPassword = settings.gets
 
         tryTimer = Time.parse(settings.gets)
         remainingTime = tryTimer - Time.now
-        puts tryTimer
-        puts remainingTime
 
         if remainingTime <= 0
+            masterPassword = settings.gets
+            
             failCount = 0
             while failCount < 5
                 puts "Please enter the Master Password:"
@@ -28,8 +32,6 @@ def startScreen
 
             if failCount >= 5
                 puts "You have tried too many times. You can try again in 20 minutes."
-                settings.rewind
-                settings.gets
                 tryTimerReset(settings)
             end
         else
@@ -37,13 +39,20 @@ def startScreen
             puts "Remaining time: #{remainingTime}s"
         end
     else
+        if File.exist?("settings")
+            puts "ERROR 002"
+            puts "Conflicting file information! Manual fix is required!"
+            exit(1)
+        end
+        
         newFile = File.new("pass", "w+")
 
         settings = File.new("settings", "w")
         puts "Please enter a Master Password:"
         masterPassword = gets
-        settings.puts(masterPassword)
+
         settings.puts(Time.now)
+        settings.puts(masterPassword)
         settings.close
 
         newFile.close
@@ -53,9 +62,7 @@ end
 def tryTimerReset(file)
     endTime = Time.now + 20*60
 
-    puts Time.now
-    puts endTime
-
+    file.rewind
     file.puts(endTime)
     file.close
 end
